// src/Server/Modules/Network.d.ts
	export namespace Common {
		export namespace Shared {
			type EnemyType = "Normal" | "Speedy" | "Tanky" | "MiniBoss" | "Boss"
			type InventoryTower = { Exp: number, Name: string, Level: number }
			type PlayerData = { Gold?: number | undefined, Silver?: number | undefined, Level?: number | undefined, Exp?: number | undefined, MaxTowers?: number | undefined, TowersInventory?: Map<string, { Exp: number, Name: string, Level: number }> | undefined, Loadout?: string[] | undefined }
		}
		export namespace Client {
			type InputAction = { Action: string, State: unknown, Key: unknown }
		}
		export namespace Server {
		}
	}
// ...
export declare namespace PlayerData {
	export namespace Common {
		export namespace Shared {
			type EnemyType = "Normal" | "Speedy" | "Tanky" | "MiniBoss" | "Boss"
			type InventoryTower = { Exp: number, Name: string, Level: number }
			type PlayerData = { Gold?: number | undefined, Silver?: number | undefined, Level?: number | undefined, Exp?: number | undefined, MaxTowers?: number | undefined, TowersInventory?: Map<string, { Exp: number, Name: string, Level: number }> | undefined, Loadout?: string[] | undefined }
		}
		export namespace Client {
			type InputAction = { Action: string, State: unknown, Key: unknown }
		}
		export namespace Server {
		}
	}
	export const replicateData: {
		fire: (Player: Player, Value: { Gold?: number | undefined, Silver?: number | undefined, Level?: number | undefined, Exp?: number | undefined, MaxTowers?: number | undefined, TowersInventory?: Map<string, { Exp: number, Name: string, Level: number }> | undefined, Loadout?: string[] | undefined }) => void
		fireAll: (Value: { Gold?: number | undefined, Silver?: number | undefined, Level?: number | undefined, Exp?: number | undefined, MaxTowers?: number | undefined, TowersInventory?: Map<string, { Exp: number, Name: string, Level: number }> | undefined, Loadout?: string[] | undefined }) => void
		fireExcept: (Except: Player, Value: { Gold?: number | undefined, Silver?: number | undefined, Level?: number | undefined, Exp?: number | undefined, MaxTowers?: number | undefined, TowersInventory?: Map<string, { Exp: number, Name: string, Level: number }> | undefined, Loadout?: string[] | undefined }) => void
		fireList: (List: Player[], Value: { Gold?: number | undefined, Silver?: number | undefined, Level?: number | undefined, Exp?: number | undefined, MaxTowers?: number | undefined, TowersInventory?: Map<string, { Exp: number, Name: string, Level: number }> | undefined, Loadout?: string[] | undefined }) => void
	}
}